Transfer donation-based views to a separate file
If you're curious about the odd name, its because there is a potential
name clash with 'donate' or 'paypal', so I just picked the first thing I
could think of. Maybe we should be prefixing the view files with an '_'.
Alternatively, we can avoid these clashes by just not having them
universally import to __init__.py, since there really isn't any point
where django relies on all the views being in the tracker.views module.
